Replace previous flask app with spots app
