+++ deepcell_label/blueprints.py
@@ -4,10 +4,13 @@ from __future__ import absolute_import, division, print_function
 import gzip
 import io
 import json
+import tempfile
 import timeit
 import traceback
 
+import boto3
 import numpy as np
+import requests
 from flask import (
     Blueprint,
     abort,
@@ -17,10 +20,10 @@ from flask import (
     request,
     send_file,
 )
-from werkzeug.exceptions import BadRequestKeyError, HTTPException
+from werkzeug.exceptions import HTTPException
 
-from deepcell_label import exporters, loaders
 from deepcell_label.label import Edit
+from deepcell_label.loaders import Loader
 from deepcell_label.models import Project
 
 bp = Blueprint('label', __name__)  # pylint: disable=C0103
@@ -30,13 +33,6 @@ bp = Blueprint('label', __name__)  # pylint: disable=C0103
 def health():
     """Returns success if the application is ready."""
     return jsonify({'message': 'success'}), 200
-
-
-@bp.errorhandler(loaders.InvalidExtension)
-def handle_invalid_extension(error):
-    response = jsonify(error.to_dict())
-    response.status_code = error.status_code
-    return response
 
 
 @bp.errorhandler(Exception)
@@ -55,73 +51,70 @@ def handle_exception(error):
     return jsonify({'error': str(error)}), 500
 
 
-# TODO: send compressed data instead of octet-stream
-@bp.route('/api/raw/<project_id>')
-def raw(project_id):
-    project = Project.get(project_id)
+@bp.route('/api/project/<project>', methods=['GET'])
+def get_project(project):
+    project = Project.get(project)
     if not project:
-        return abort(404, description=f'project {project_id} not found')
-    # Send binary data for raw image array
-    raw = project.raw_array
-    raw = (
-        (raw - np.min(raw, axis=(0, 1, 2)))
-        / (np.max(raw, axis=(0, 1, 2)) - np.min(raw, axis=(0, 1, 2)))
-        * 255
+        return abort(404, description=f'project {project} not found')
+    s3 = boto3.client('s3')
+    data = io.BytesIO()
+    s3.download_fileobj(project.bucket, project.key, data)
+    data.seek(0)
+    return send_file(data, mimetype='application/zip')
+
+
+@bp.route('/api/project', methods=['POST'])
+def create_project():
+    """
+    Create a new Project from URL.
+    """
+    start = timeit.default_timer()
+    images_url = request.form['images']
+    labels_url = request.form['labels'] if 'labels' in request.form else None
+    # dimension_order = (
+    #     request.form['dimension_order']
+    #     if 'dimension_order' in request.form
+    #     else 'TZYXC'
+    # )
+    # labels_dimension_order = (
+    #     request.form['labels_dimension_order']
+    #     if 'labels_dimension_order' in request.form
+    #     else None
+    # )
+    with tempfile.TemporaryFile() as image_file, tempfile.TemporaryFile() as label_file:
+        image_file.write(requests.get(images_url).content)
+        image_file.seek(0)
+        if labels_url is not None:
+            label_file.write(requests.get(labels_url).content)
+            label_file.seek(0)
+        loader = Loader(image_file, None if labels_url is None else label_file)
+        project = Project.create(loader.data)
+    current_app.logger.info(
+        'Created project %s from %s in %s s.',
+        project.project,
+        f'{images_url}' if labels_url is None else f'{images_url} and {labels_url}',
+        timeit.default_timer() - start,
     )
-    # Reshape (frames, height, width, channels) to (channels, frames, height, width)
-    raw = np.moveaxis(raw, -1, 0)
-    raw = raw.astype('uint8')
-    return send_file(io.BytesIO(raw.tobytes()), mimetype='application/octet-stream')
+    return jsonify(project.project)
 
 
-@bp.route('/api/labeled/<project_id>')
-def labeled(project_id):
-    project = Project.get(project_id)
-    if not project:
-        return abort(404, description=f'project {project_id} not found')
-    # send binary data for label array (int16? int32?)
-    labeled = project.label_array
-    # Reshape (frames, height, width, features) to (features, frames, height, width)
-    labeled = np.moveaxis(labeled, -1, 0)
-    labeled = labeled.astype('int32')
-    return send_file(io.BytesIO(labeled.tobytes()), mimetype='application/octet-stream')
-
-
-@bp.route('/api/labels/<project_id>')
-def labels(project_id):
-    project = Project.get(project_id)
-    if not project:
-        return abort(404, description=f'project {project_id} not found')
-    # send JSON data
-    labels = project.labels.cell_info
-    return labels
-
-
-@bp.route('/api/raw/<token>', methods=['POST'])
-def add_raw(token):
-    """Add new channel to the project."""
-    project = Project.get(token)
-    if not project:
-        return abort(404, description=f'project {token} not found')
-    # Load channel from first array in attached file
-    try:
-        npz = np.load(request.files.get('file'))
-    except BadRequestKeyError:  # could not get file from request.files
-        return abort(
-            400, description='Attach a new channel file in a form under the file field.'
-        )
-    except TypeError:
-        return abort(
-            400, description='Could not load the attached file. Attach an .npz file.'
-        )
-    channel = npz[npz.files[0]]
-    # Check channel is the right shape
-    expected_shape = (project.num_frames, project.width, project.height, 1)
-    if channel.shape != expected_shape:
-        raise ValueError(f'New channel must have shape {expected_shape}')
-    # Add channel to project
-    project.add_channel(channel)
-    return {'numChannels': project.num_channels}
+@bp.route('/api/project/dropped', methods=['POST'])
+def create_project_from_dropped_file():
+    """
+    Create a new Project from drag & dropped file.
+    """
+    start = timeit.default_timer()
+    input_file = request.files.get('file')
+    # axes = request.form['axes'] if 'axes' in request.form else DCL_AXES
+    loader = Loader(input_file)
+    project = Project.create(loader.data)
+    current_app.logger.info(
+        'Created project %s from %s in %s s.',
+        project.project,
+        input_file.filename,
+        timeit.default_timer() - start,
+    )
+    return jsonify(project.project)
 
 
 @bp.route('/api/edit/<action>', methods=['POST'])
@@ -168,98 +161,6 @@ def edit(action):
     return response
 
 
-@bp.route('/api/undo/<token>', methods=['POST'])
-def undo(token):
-    start = timeit.default_timer()
-
-    project = Project.get(token)
-    if not project:
-        return abort(404, description=f'project {token} not found')
-    payload = project.undo()
-
-    current_app.logger.debug(
-        'Undid action for project %s finished in %s s.',
-        token,
-        timeit.default_timer() - start,
-    )
-    return jsonify(payload)
-
-
-@bp.route('/api/redo/<token>', methods=['POST'])
-def redo(token):
-    start = timeit.default_timer()
-
-    project = Project.get(token)
-    if not project:
-        return abort(404, description=f'project {token} not found')
-    payload = project.redo()
-
-    current_app.logger.debug(
-        'Redid action for project %s finished in %s s.',
-        token,
-        timeit.default_timer() - start,
-    )
-    return jsonify(payload)
-
-
-@bp.route('/api/project/<token>', methods=['GET'])
-def get_project(token):
-    """
-    Retrieve data from a project already in the Project table.
-    """
-    start = timeit.default_timer()
-    project = Project.get(token)
-    if not project:
-        return abort(404, description=f'project {token} not found')
-    payload = project.make_first_payload()
-    current_app.logger.debug(
-        'Loaded project %s in %s s.', project.token, timeit.default_timer() - start
-    )
-    return jsonify(payload)
-
-
-# @bp.route('/api/project', methods=['POST'])
-# def create_project():
-#     """
-#     Create a new Project.
-#     """
-#     start = timeit.default_timer()
-#     loader = loaders.get_loader(request)
-#     project = Project.create(loader)
-#     current_app.logger.info('Created project from %s in %s s.',
-#                             loader.path, timeit.default_timer() - start)
-#     return jsonify({'projectId': project.token})
-
-
-@bp.route('/api/project/dropped', methods=['POST'])
-def create_project_from_dropped_file():
-    """
-    Create a new Project from drag & dropped file.
-    """
-    start = timeit.default_timer()
-    loader = loaders.FileLoader(request)
-    project = Project.create(loader)
-    current_app.logger.info(
-        'Created project from %s in %s s.', loader.path, timeit.default_timer() - start
-    )
-    return jsonify({'projectId': project.token})
-
-
-@bp.route('/api/project', methods=['POST'])
-def create_project_from_url():
-    """
-    Create a new Project from URL.
-    """
-    start = timeit.default_timer()
-    url_form = request.form
-    loader = loaders.URLLoader(url_form)
-    project = Project.create(loader)
-    current_app.logger.info(
-        'Created project from %s in %s s.', loader.path, timeit.default_timer() - start
-    )
-    return jsonify({'projectId': project.token})
-
-
 @bp.route('/api/download', methods=['GET'])
 def download_project():
     """
@@ -269,35 +170,8 @@ def download_project():
     project = Project.get(id)
     if not project:
         return abort(404, description=f'project {id} not found')
-    format = request.args.get('format')
-    exporter = exporters.Exporter(project, format)
-    filestream = exporter.export()
-
-    return send_file(filestream, as_attachment=True, attachment_filename=exporter.path)
-
-
-@bp.route('/api/upload', methods=['GET', 'POST'])
-def upload_project_to_s3():
-    """Upload .trk/.npz data file to AWS S3 bucket."""
-    start = timeit.default_timer()
-    id = request.form['id']
-    format = request.form['format']
-    bucket = request.form['bucket']
-    project = Project.get(id)
-    if not project:
-        return abort(404, description=f'project {id} not found')
-
-    # Save data file and send to S3 bucket
-    exporter = exporters.S3Exporter(project, format)
-    exporter.export(bucket)
-    # add "finished" timestamp and null out PickleType columns
-    # project.finish()
-
-    current_app.logger.debug(
-        'Uploaded %s to S3 bucket %s from project %s in %s s.',
-        exporter.path,
-        bucket,
-        id,
-        timeit.default_timer() - start,
-    )
-    return {}
+    s3 = boto3.client('s3')
+    data = io.BytesIO()
+    s3.download_fileobj(project.bucket, project.key, data)
+    data.seek(0)
+    return send_file(data, as_attachment=True, attachment_filename=project.key)
