+++ deepcell_label/loaders.py
@@ -1,357 +1,233 @@
 """
-Classes to load external data from URLs into a DeepCell Label Project.
-Loads or creates raw_array, label_array, cell_ids, and cell_info.
+Class to load data into a DeepCell Label project file
+Loads both raw image data and labels
 """
 
 import io
 import json
-import pathlib
-import tarfile
 import tempfile
 import zipfile
 
+import magic
 import numpy as np
-import requests
 from PIL import Image
-from tifffile import TiffFile
+from tifffile import TiffFile, TiffWriter
 
 from deepcell_label.labelmaker import LabelInfoMaker
-from deepcell_label.utils import reshape
-
-DCL_AXES = 'ZYXC'
 
 
 class Loader:
     """
-    Interface for loading files into DeepCell Label.
+    Loads data into a DeepCell Label project file
     """
 
-    def __init__(self):
-        self.path = ''
-        self.labeled_path = ''
-        self.raw_array = None
-        self.label_array = None
-        self.cell_info = None
-        self.cell_ids = None
+    def __init__(self, image_file, label_file=None):
+        self._image_file = image_file
+        self._label_file = image_file if label_file is None else label_file
+        with tempfile.TemporaryFile() as project_file:
+            with zipfile.ZipFile(project_file, 'w', zipfile.ZIP_DEFLATED) as zip:
+                self.zip = zip
+                self.load()
+            project_file.seek(0)
+            self.data = project_file.read()
 
     @property
-    def is_tracking(self):
-        """Tracking project when either the labels or raw are .trk."""
-        return (self.labeled_path and is_trk(self.labeled_path)) or is_trk(self.path)
+    def image_file(self):
+        self._image_file.seek(0)
+        return self._image_file
 
-    def add_semantic_labels(self):
-        label_maker = LabelInfoMaker(self.label_array, tracking=True)
-        self.cell_ids = label_maker.cell_ids
-        if self.cell_info is None:
-            self.cell_info = label_maker.cell_info
-
-    def load_combined(self, data):
-        """
-        Loads image data into the Loader based on the file extension.
-        """
-        label_array = None
-        # Load arrays
-        if is_npz(self.path):
-            raw_array = load_raw_npz(data)
-            label_array = load_labeled_npz(data)
-        elif is_trk(self.path):
-            raw_array = load_raw_trk(data)
-            label_array = load_labeled_trk(data)
-        elif is_png(self.path):
-            raw_array = load_png(data)
-        elif is_tiff(self.path):
-            raw_array = load_tiff(data)
-        elif is_zip(self.path):
-            raw_array = load_zip(data)
-        else:
-            ext = pathlib.Path(self.path).suffix
-            raise InvalidExtension('invalid file extension: {}'.format(ext))
-
-        # Reshape or create arrays
-        raw_array = reshape(raw_array, self.input_axes, self.output_axes)
-        if label_array is None:
-            # Substitute channels dimension with one feature
-            shape = (*raw_array.shape[:-1], 1)
-            label_array = np.zeros(shape)
-        else:
-            label_array = reshape(label_array, self.input_axes, self.output_axes)
-
-        self.raw_array = raw_array
-        self.label_array = label_array
-
-    def load_raw(self, data):
-        if is_npz(self.path):
-            raw_array = load_raw_npz(data)
-        elif is_trk(self.path):
-            raw_array = load_raw_trk(data)
-        elif is_png(self.path):
-            raw_array = load_png(data)
-        elif is_tiff(self.path):
-            raw_array = load_tiff(data)
-        elif is_zip(self.path):
-            raw_array = load_zip(data)
-        else:
-            ext = pathlib.Path(self.path).suffix
-            raise InvalidExtension('invalid file extension: {}'.format(ext))
-
-        self.raw_array = reshape(raw_array, self.input_axes, self.output_axes)
-
-    def load_labeled(self, data):
-        if is_npz(self.labeled_path):
-            label_array = load_labeled_npz(data)
-            if label_array is None:
-                label_array = load_npz(data)
-        elif is_trk(self.labeled_path):
-            label_array = load_labeled_trk(data)
-        elif is_png(self.labeled_path):
-            label_array = load_png(data)
-        elif is_tiff(self.labeled_path):
-            label_array = load_tiff(data)
-        elif is_zip(self.labeled_path):
-            label_array = load_zip(data)
-        else:
-            ext = pathlib.Path(self.labeled_path).suffix
-            raise InvalidExtension('invalid file extension: {}'.format(ext))
-
-        self.label_array = reshape(label_array, 'CZYX', self.output_axes)
-
-
-class FileLoader(Loader):
-    """
-    Loader implementation for files sent in a request, like a drag-and-dropped file.
-    """
-
-    def __init__(self, request):
-        super(FileLoader, self).__init__()
-        file_ = request.files.get('file')
-        self.input_axes = request.form['axes'] if 'axes' in request.form else DCL_AXES
-        self.output_axes = DCL_AXES
-        self.data = file_.stream.read()
-        self.path = file_.filename
-        self.source = 'dropped'
-        self.load()
+    @property
+    def label_file(self):
+        self._label_file.seek(0)
+        return self._label_file
 
     def load(self):
-        label_array = None
-        # Load arrays
-        if is_npz(self.path):
-            raw_array = load_raw_npz(self.data)
-            label_array = load_labeled_npz(self.data)
-        elif is_png(self.path):
-            raw_array = load_png(self.data)
-        elif is_tiff(self.path):
-            raw_array = load_tiff(self.data)
-        elif is_trk(self.path):
-            raw_array = load_raw_trk(self.data)
-            label_array = load_labeled_trk(self.data)
-            self.cell_info = load_lineage_trk(self.data)
+        """Extracts data from input files and writes standardized data to output zip."""
+        # Load data from input files
+        X = self.load_images()
+        y = self.load_segmentation()
+        spots = self.load_spots()
+
+        # Write standardized data to output zip
+        if X is not None:
+            # Rescale data before casting to uint8
+            X = (X - np.min(X)) / (np.max(X) - np.min(X)) * 255
+            X = X.astype(np.uint8)
+            self.write_images(X)
         else:
-            ext = pathlib.Path(self.path).suffix
-            raise InvalidExtension('Invalid file extension: {}'.format(ext))
+            raise ValueError('No images found in files')
 
-        # Reshape or create arrays
-        raw_array = reshape(raw_array, self.input_axes, self.output_axes)
-        if label_array is None:
-            # Substitute channels dimension with one feature
-            shape = (*raw_array.shape[:-1], 1)
-            label_array = np.zeros(shape)
+        if y is not None:
+            if y.shape[:-1] != X.shape[:-1]:
+                raise ValueError(
+                    'Segmentation shape %s is incompatible with image shape %s'
+                    % (y.shape, X.shape)
+                )
+            # TODO: check if float vs int matters
+            self.write_segmentation(y.astype(np.uint32))
+            # Write cells in segmentation
+            cells = LabelInfoMaker(y).cell_info
+            self.write_cells(cells)
+
+        if spots is not None:
+            self.write_spots(spots)
+
+    def write_images(self, X):
+        """Writes raw images to output zip."""
+        images = io.BytesIO()
+        with TiffWriter(images, ome=True) as tif:
+            tif.save(np.moveaxis(X, -1, 0), metadata={'axes': 'CZYX'})
+        images.seek(0)
+        self.zip.writestr('X.ome.tiff', images.read())
+        print('Wrote images')
+
+    def write_segmentation(self, y):
+        """Writes segmentation to output zip."""
+        segmentation = io.BytesIO()
+        with TiffWriter(segmentation, ome=True) as tif:
+            tif.save(np.moveaxis(y, -1, 0), metadata={'axes': 'CZYX'})
+        segmentation.seek(0)
+        self.zip.writestr('y.ome.tiff', segmentation.read())
+        print('Wrote segmentation')
+
+    def write_spots(self, spots):
+        """Writes spots to output zip."""
+        buffer = io.BytesIO()
+        buffer.write(spots)
+        buffer.seek(0)
+        self.zip.writestr('spots.csv', buffer.read())
+        print('Wrote spots')
+
+    def write_cells(self, cells):
+        """Writes cells.json to output zip."""
+        self.zip.writestr('cells.json', json.dumps(cells))
+        print('Wrote cells')
+
+    def load_images(self):
+        """Extracts raw images from input image file."""
+        X = load_zip(self.image_file)
+        if X is None:
+            X = load_npy(self.image_file)
+        if X is None:
+            X = load_tiff(self.image_file)
+        if X is None:
+            X = load_png(self.image_file)
+        return X
+
+    def load_segmentation(self):
+        """Extracts segmentation from input label file."""
+        if zipfile.is_zipfile(self.label_file):
+            label_zip = zipfile.ZipFile(self.label_file, 'r')
+            y = load_zip_numpy(label_zip, name='y')
+            if y is None:
+                y = load_zip_tiffs(label_zip)
+            return y
+
+    def load_spots(self):
+        """Extracts spots from input label file."""
+        if zipfile.is_zipfile(self.label_file):
+            label_zip = zipfile.ZipFile(self.label_file, 'r')
+            return load_zip_csv(label_zip)
+
+
+def load_zip_numpy(zf, name='X'):
+    """
+    Loads a numpy array from the zip file, if it exists.
+    If loading from an NPZ with multiple arrays, use name parameter to pick one.
+    """
+    for filename in zf.namelist():
+        if filename == f'{name}.npy':
+            with zf.open(filename) as f:
+                return np.load(f)
+        if filename.endswith('.npz'):
+            with zf.open(filename) as f:
+                npz = np.load(f)
+                return npz[name] if name in npz.files else npz[npz.files[0]]
+
+
+def load_zip_tiffs(zf):
+    """
+    Returns an array with all tiff image data in the zip file, if any.
+    """
+    tiffs = []
+    for name in zf.namelist():
+        with zf.open(name) as f:
+            if 'TIFF image data' in magic.from_buffer(f.read(2048)):
+                f.seek(0)
+                tiff = TiffFile(f).asarray()
+                tiffs.append(tiff)
+    if len(tiffs) > 0:
+        # Stack channels on last axis
+        y = np.stack(tiffs, axis=-1)
+        # Add frame axis
+        if y.ndim == 3:
+            y = y[np.newaxis, ...]
+        return y
+
+
+def load_zip_png(zf):
+    """
+    Returns the image data array for the first PNG image in the zip file, if it exists.
+    """
+    for name in zf.namelist():
+        with zf.open(name) as f:
+            if 'PNG image data' in magic.from_buffer(f.read(2048)):
+                f.seek(0)
+                png = Image.open(f)
+                return np.array(png)
+
+
+def load_zip_csv(z):
+    """
+    Returns the binary data for the first CSV file in the zip file, if it exists.
+    """
+    for name in z.namelist():
+        with z.open(name) as f:
+            if 'CSV text' in magic.from_buffer(f.read(2048)):
+                f.seek(0)
+                return f.read()
+
+
+def load_zip(f):
+    """Loads image data from a zip file by loading from the npz, tiff, or png files in the archive."""
+    if zipfile.is_zipfile(f):
+        zf = zipfile.ZipFile(f, 'r')
+        X = load_zip_numpy(zf)
+        if X is None:
+            X = load_zip_tiffs(zf)
+        if X is None:
+            X = load_zip_png(zf)
+        return X
+
+
+def load_npy(f):
+    """Loads image data from a npy file"""
+    if 'NumPy data file' in magic.from_buffer(f.read(2048)):
+        f.seek(0)
+        npy = np.load(f)
+        return npy
+
+
+def load_tiff(f):
+    """Loads image data from a tiff file"""
+    if 'TIFF image data' in magic.from_buffer(f.read(2048)):
+        f.seek(0)
+        return TiffFile(f).asarray()
+
+
+def load_png(f):
+    """Loads image data from a png file"""
+    if 'PNG image data' in magic.from_buffer(f.read(2048)):
+        f.seek(0)
+        image = Image.open(f, formats=['PNG'])
+        # Luminance should add channel dimension at end
+        if image.mode == 'L':
+            X = np.array(image)
+            X = np.expand_dims(X, -1)
         else:
-            label_array = reshape(label_array, self.input_axes, self.output_axes)
-
-        self.raw_array = raw_array
-        self.label_array = label_array
-
-        self.add_semantic_labels()
-
-
-class URLLoader(Loader):
-    """
-    Loader implementation for downloading files from URLs.
-    """
-
-    def __init__(self, url_form):
-        super(URLLoader, self).__init__()
-        self.source = 's3'
-        self.path = url_form['url']
-        self.labeled_path = (
-            url_form['labeled_url'] if 'labeled_url' in url_form else None
-        )
-        self.input_axes = url_form['axes'] if 'axes' in url_form else DCL_AXES
-        self.output_axes = DCL_AXES
-
-        self.load()
-
-    def load(self):
-        data = requests.get(self.path).content
-        if self.labeled_path is None:
-            self.load_combined(data)
-            if is_trk(self.path):
-                self.cell_info = load_lineage_trk(data)
-        else:
-            labeled_data = requests.get(self.labeled_path).content
-            self.load_raw(data)
-            self.load_labeled(labeled_data)
-            if is_trk(self.labeled_path):
-                self.cell_info = load_lineage_trk(labeled_data)
-
-        self.add_semantic_labels()
-
-
-def is_npz(url):
-    return pathlib.Path(url).suffix in {'.npz'}
-
-
-def is_trk(url):
-    return pathlib.Path(url).suffix in {'.trk', '.trks'}
-
-
-def is_png(url):
-    return pathlib.Path(url).suffix in {'.png'}
-
-
-def is_tiff(url):
-    return pathlib.Path(url).suffix in {'.tiff', '.tif'}
-
-
-def is_zip(url):
-    return pathlib.Path(url).suffix in {'.zip'}
-
-
-def load_npz(data):
-    """Returns the first array in an npz."""
-    npz = np.load(io.BytesIO(data))
-    return npz[npz.files[0]]
-
-
-def load_raw_npz(data):
-    """
-    Returns raw image array from an NPZ file.
-    """
-    npz = np.load(io.BytesIO(data))
-
-    # standard names for image (X) and labeled (y)
-    if 'X' in npz.files:
-        return npz['X']
-    # alternate names 'raw' and 'annotated'
-    elif 'raw' in npz.files:
-        return npz['raw']
-    # if filenames are different, try to load them anyways
-    else:
-        return npz[npz.files[0]]
-
-
-def load_labeled_npz(data):
-    """
-    Returns labeled image array from an NPZ file.
-    Returns None when the labeled image array is not present.
-    """
-    npz = np.load(io.BytesIO(data))
-
-    # Look for label filenames
-    if 'y' in npz.files:
-        return npz['y']
-    elif 'annotated' in npz.files:
-        return npz['annotated']
-    elif len(npz.files) > 1:
-        return npz[npz.files[1]]
-
-
-def load_raw_trk(data):
-    """Load a raw image data from a .trk file."""
-    with tempfile.NamedTemporaryFile() as temp:
-        temp.write(data)
-        with tarfile.open(temp.name, 'r') as trks:
-            with io.BytesIO() as array_file:
-                array_file.write(trks.extractfile('raw.npy').read())
-                array_file.seek(0)
-                return np.load(array_file)
-
-
-def load_labeled_trk(data):
-    """Load a labeled image data from a .trk file."""
-    with tempfile.NamedTemporaryFile() as temp:
-        temp.write(data)
-        with tarfile.open(temp.name, 'r') as trks:
-            with io.BytesIO() as array_file:
-                array_file.write(trks.extractfile('tracked.npy').read())
-                array_file.seek(0)
-                return np.load(array_file)
-
-
-def load_lineage_trk(data):
-    """Loads a lineage JSON from a .trk file."""
-    with tempfile.NamedTemporaryFile() as temp:
-        temp.write(data)
-        with tarfile.open(temp.name, 'r') as trks:
-            try:
-                trk_data = trks.getmember('lineages.json')
-            except KeyError:
-                try:
-                    trk_data = trks.getmember('lineage.json')
-                except KeyError:
-                    raise ValueError('Invalid .trk file, no lineage data found.')
-
-            lineages = json.loads(trks.extractfile(trk_data).read().decode())
-            lineages = lineages if isinstance(lineages, list) else [lineages]
-
-            # JSON only allows strings as keys, so convert them back to ints
-            for i, tracks in enumerate(lineages):
-                lineages[i] = {int(k): v for k, v in tracks.items()}
-
-            # Track files have only one feature and one lineage
-            if len(lineages) != 1:
-                raise ValueError('Input file has multiple trials/lineages.')
-            return {0: lineages[0]}
-
-
-def load_png(data):
-    """Returns image array from a PNG file."""
-    image = Image.open(io.BytesIO(data), formats=['PNG'])
-    # Luminance should add channel dimension at end
-    if image.mode == 'L':
-        array = np.array(image)
-        array = np.expand_dims(array, -1)
-    else:
-        # Create three RGB channels
-        # Handles RGB, RGBA, P modes
-        array = np.array(image.convert('RGB'))
-    # Add frame dimension at start
-    array = np.expand_dims(array, 0)
-    return array
-
-
-def load_tiff(data):
-    """Returns image array from a TIFF file."""
-    # return np.asarray(imageio.imread(data))
-    return TiffFile(io.BytesIO(data)).asarray()
-
-
-def load_zip(data):
-    """
-    Loads a series of image arrays from a zip of TIFFs.
-    Treats separate TIFFs as channels.
-    """
-    zip_file = zipfile.ZipFile(io.BytesIO(data), 'r')
-    channels = [
-        load_tiff(zip_file.open(item).read())
-        for item in zip_file.filelist
-        if not str(item.filename).startswith('__MACOSX/')
-        and is_tiff(str(item.filename))
-    ]
-    return np.array(channels)
-
-
-class InvalidExtension(Exception):
-    status_code = 415
-
-    def __init__(self, message, status_code=None, payload=None):
-        Exception.__init__(self)
-        self.message = message
-        if status_code is not None:
-            self.status_code = status_code
-        self.payload = payload
-
-    def to_dict(self):
-        rv = dict(self.payload or ())
-        rv['error'] = self.message
-        return rv
+            # Create three RGB channels
+            # Handles RGB, RGBA, P modes
+            X = np.array(image.convert('RGB'))
+        # Add frame dimension at start
+        X = np.expand_dims(X, 0)
+        return X
